+++ R/stages_bhc.R
@@ -93,7 +93,7 @@ stages_bhc <-
     if (trace > 0) {
       cli::cli_text("BHC done")
     }
-    object$call <- match.call()
+    object <- record_call(object, match.call())
     object$score <- list(value = now_score, f = score)
     return(object)
   }