+++ R/stages_bj.R
@@ -101,6 +101,6 @@ stages_bj <-
     if (trace > 0) {
       cli::cli_text("backward join done.")
     }
-    object$call <- match.call()
+    object <- record_call(object, match.call())
     return(object)
   }